requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
langchain-ollama
langchain-nomic
groq
langchain-huggingface
transformers
python-dotenv
numpy
scikit-learn
chromadb
pypdf
youtube-transcript-api
faiss-cpu
tiktoken
duckduckgo-search
ddgs
langgraph
requests
langgraph-checkpoint-sqlite
langchain-chroma